Refactor: refactor main and login page 메인, 로그인 페이지 커스텀 훅 생성 다른 파일들 파일 위치 및 이름 변경
