+++ alcoholmap/src/frontend/pages/index.tsx
@@ -1,4 +1,4 @@
-import MainPage from "../src/components/units/main/Main"
+import MainPage from "../src/components/units/main/Main.container"
 
 export default function Home():JSX.Element{
     return <MainPage/>
