+++ alcoholmap/src/frontend/pages/map/index.tsx
@@ -1,4 +1,4 @@
-import MapPage from '../../src/components/units/map/Map.container'
+import MapPage from '../../src/components/units/map/mapmain/Map.container'
 
 export default function Map():JSX.Element{
     return <MapPage/>
