+++ alcoholmap/src/frontend/src/components/units/login/Login.container.tsx
@@ -1,20 +1,13 @@
-import { useState } from 'react'
-import { useRouter } from 'next/router'
-import LoginUI from './Login.presenter'
-import axios from 'axios';
-import { LoginForm } from './Login.types';
+import { useRouter } from 'next/router';
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { loginSchema } from "../../../commons/yupSchemas";
-
-/*  백엔드 서버에 이메일, 비밀번호 보내기
-    로그인 성공 시 메인페이지 라우터
-*/
-const apiUrl = '/users/login';
+import LoginUI from './Login.presenter';
+import { LoginForm } from './Login.types';
+import useLogin from '../../../hooks/login/useLogin';
+import { loginSchema } from "../../../commons/util/yupSchemas";
 
 export default function LoginPage(): JSX.Element{
     const router = useRouter()
-
     const { register, handleSubmit, formState } = useForm<LoginForm>({
         mode: 'onChange',
         resolver: yupResolver(loginSchema),
@@ -25,43 +18,22 @@ export default function LoginPage(): JSX.Element{
         },
         shouldFocusError: true,
         shouldUnregister: true,
-      });
-    const [jwtToken, setJwtToken] = useState("");
-    //const [disabledbutton, setButton] = useState(true) //버튼 비활성화 추후 추가
+    });
+
+    const { onSendLoginForm } = useLogin();
 
     const onSubmit: SubmitHandler<LoginForm> = (data: LoginForm) => {
-        // 로그인 처리 로직 추가
-        console.log(data);
         onSendLoginForm(data);
-      };
-
-    const onSendLoginForm = async (loginForm: LoginForm) => {
-        const loginFormJson = JSON.stringify(loginForm);
-        console.log(loginFormJson);
-        try {
-            const response = await axios.post(apiUrl, loginFormJson, {
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-            });
-            const token = response.data.accessToken;
-            //const token = 'eyJhbGciOiJIUzI1NiJ9.eyJlbWFpbCI6ImtpbXRheDBAZ21haWwuY29tIiwicGFzc3dvcmQiOiJraW10YXgwMTIzISJ9.E67z1F14tAT1Yz7DeU6LLlBYXLkzHoP8k7qumkn3DgA';
-            localStorage.setItem('jwtToken', token);
-            router.push("../map"); //로그인 성공 시 메인페이지로 이동
-        } catch (error){
-            console.error('error submitting data:', error);
-            alert("로그인에 실패했습니다. 이메일과 비밀번호를 확인해주세요.")
-        }
     };
 
     const onClickMoveToSignup = (): void => {
         router.push("../signup")
-    }
+    };
 
     const onClickMoveToMainpage = (): void => {
         router.push("../map")
-    }
-    console.log("rerendering")
+    };
+
     return (
         <LoginUI
             formMethods={{ register, handleSubmit, formState }}
@@ -69,5 +41,5 @@ export default function LoginPage(): JSX.Element{
             onClickMoveToSignup = {onClickMoveToSignup}
             onClickMoveToMainpage = {onClickMoveToMainpage}
         />
-    )
+    );
 }
