+++ alcoholmap/src/frontend/src/components/units/main/Main.types.ts
@@ -1,0 +1,12 @@
+export interface IMainUIProps {
+    mobImages:string[];
+    webImages:string[];
+    currentImageIndex:number;
+    isFlashing:boolean;
+    isTransitioning:boolean;
+    onClickMoveToMainpage: () => void;
+    onClickMoveToLogin: () => void;
+    onClickMoveToSignup: () => void;
+    prevSlide: () => void;
+    nextSlide: () => void;
+}
