+++ alcoholmap/src/frontend/src/components/units/map/maphbmenu/MapHBMenu.types.ts
