+++ alcoholmap/src/frontend/src/components/units/map/mapmain/Map.presenter.tsx
@@ -1,8 +1,8 @@
 import * as S from "./Map.styles";
 import { IMapUIProps } from "./Map.types";
 import { useState, useRef} from "react";
-import MapHBMenu from "./MapHBMenu";
-import MapHelp from "./MapHelp";
+import MapHBMenu from "../maphbmenu/MapHBMenu";
+import MapHelp from "../maphelp/MapHelp";
 
 export default function MapUI(props: IMapUIProps): JSX.Element{
     const [isPlaceListOpen, setIsPlaceListOpen] = useState(true);
