+++ alcoholmap/src/frontend/src/components/units/map/mapmain/Map.types.ts
