+++ alcoholmap/src/frontend/src/components/units/map/modal/Modal.types.ts
