+++ alcoholmap/src/frontend/src/components/units/signup/Signup.container.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/router'
 import { SignupForm } from './Signup.types';
 import { useForm, SubmitHandler } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
-import { signupSchema } from "../../../commons/yupSchemas";
+import { signupSchema } from "../../../commons/util/yupSchemas";
 
 /*  백엔드 서버에 이메일아이디 + @ + 도메인 합쳐서 보내기
     비밀번호 보내기
